Beta Update
Improved stealth by using a DEF file to ensure the name in the Exports
section matches, as well as to ensure the ordinals of the exports in the
proxy DLL match the ordinals of the original.

This updated includes many bug fixes through a much more therough
testing of DLL Export Functionality.

Configuration File handling needs to continue to be tested before final
release.